# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Contacts API",
    "version": "1.0.0",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "API for managing contacts with user authentication, password recovery, and image uploads."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://nodejs-hw-mongodb-w4df.onrender.com"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "User authentication endpoints"
    },
    {
      "name": "Contacts",
      "description": "Endpoints for managing user contacts"
    }
  ],
  "paths": {
    "/users/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "operationId": "registerUser",
        "description": "Registers a new user with a unique email and password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Test User"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "test@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "password123"
                  }
                },
                "required": [
                  "name",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/registeredUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., validation error)"
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get all contacts",
        "operationId": "getAllContacts",
        "description": "Retrieve a list of all contacts for the authenticated user. Supports pagination.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Number of contacts per page"
          },
          {
            "in": "query",
            "name": "favorite",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter contacts by favorite status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with a list of contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contacts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/contact"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "Total number of contacts"
                    },
                    "page": {
                      "type": "integer",
                      "description": "Current page number"
                    },
                    "limit": {
                      "type": "integer",
                      "description": "Number of contacts per page"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create a new contact",
        "operationId": "createContact",
        "description": "Create a new contact for the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/contact"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/contact"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., validation error)"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          }
        }
      }
    },
    "/contacts/{contactId}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get a contact by ID",
        "operationId": "getContactById",
        "description": "Retrieve a single contact for the authenticated user by its ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "ID of the contact to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with the contact details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/contact"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Contacts"
        ],
        "summary": "Update a contact by ID",
        "operationId": "updateContactById",
        "description": "Update an existing contact for the authenticated user by its ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "ID of the contact to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/contact"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/contact"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., validation error)"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete a contact by ID",
        "operationId": "deleteContactById",
        "description": "Delete a contact for the authenticated user by its ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "ID of the contact to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Contact successfully deleted"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Contact": {
        "$ref": "#/components/schemas/contact"
      },
      "RegisteredUserResponse": {
        "$ref": "#/components/schemas/registeredUserResponse"
      },
      "PaginatedContacts": {
        "$ref": "#/components/schemas/paginatedContacts"
      },
      "registeredUserResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 201,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "Successfully registered a user!",
            "description": "A message indicating the success of the registration."
          },
          "data": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Test User"
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "test@example.com"
              },
              "_id": {
                "type": "string",
                "pattern": "^[0-9a-fA-F]{24}$",
                "description": "Unique identifier of the user",
                "example": "686ed7bbaf6ab84c5a3978b0"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "description": "Date and time when the user was created"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time",
                "description": "Date and time when the user was last updated"
              },
              "__v": {
                "type": "integer",
                "example": 0
              }
            },
            "required": [
              "name",
              "email",
              "_id",
              "createdAt",
              "updatedAt"
            ]
          }
        }
      },
      "contact": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique identifier of the contact",
            "example": "60d5ec49f3e4e9001c8c4a4a"
          },
          "name": {
            "type": "string",
            "description": "Name of the contact",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the contact",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the contact",
            "example": "123-456-7890"
          },
          "favorite": {
            "type": "boolean",
            "description": "Indicates if the contact is a favorite",
            "example": false
          },
          "owner": {
            "type": "string",
            "description": "ID of the user who owns the contact",
            "example": "60d5ec49f3e4e9001c8c4a4b"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the contact was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the contact was last updated"
          }
        }
      },
      "paginatedContacts": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "example": 1,
                "description": "Current page number"
              },
              "limit": {
                "type": "integer",
                "example": 10,
                "description": "Number of items per page"
              },
              "totalPages": {
                "type": "integer",
                "example": 5,
                "description": "Total number of pages"
              },
              "totalCount": {
                "type": "integer",
                "example": 42,
                "description": "Total number of items"
              },
              "contacts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/contact"
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "NotFound": {
        "$ref": "#/components/responses/notFound"
      },
      "Unauthorized": {
        "$ref": "#/components/responses/unauthorized"
      },
      "unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Not authorized"
                }
              }
            }
          }
        }
      },
      "notFound": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Not found"
                }
              }
            }
          }
        }
      }
    }
  }
}